use smaller data type for oled_write_pixel

+++ drivers/oled/oled_driver.h
@@ -208,7 +208,7 @@ void oled_write_raw_byte(const char data, uint16_t index);
 
 // Sets a specific pixel on or off
 // Coordinates start at top-left and go right and down for positive x and y
-void oled_write_pixel(int16_t x, int16_t y, bool on);
+void oled_write_pixel(uint8_t x, uint8_t y, bool on);
 
 #if defined(__AVR__)
 // Writes a PROGMEM string to the buffer at current cursor position
